+++ src/compas_timber/model/model.py
@@ -12,8 +12,8 @@ from compas.tolerance import TOL
 from compas_model.models import Model
 
 from compas_timber.connections import ConnectionSolver
-from compas_timber.connections import JointCandidate
 from compas_timber.connections import Joint
+from compas_timber.connections import JointCandidate
 from compas_timber.connections import JointTopology
 from compas_timber.connections import WallJoint
 from compas_timber.errors import BeamJoiningError
@@ -30,8 +30,10 @@ class TimberModel(Model):
         A Generator object of all beams assigned to this model.
     plates : Generator[:class:`~compas_timber.elements.Plate`]
         A Generator object of all plates assigned to this model.
-    joints : Generator[:class:`~compas_timber.connections.Joint`]
-        A Generator object of all joints assigned to this model.
+    joints : set[:class:`~compas_timber.connections.Joint`]
+        A set of all actual joints assigned to this model.
+    joint_candidates : set[:class:`~compas_timber.connections.JointCandidate`]
+        A set of all joint candidates in the model.
     walls : Generator[:class:`~compas_timber.elements.Wall`]
         A Generator object of all walls assigned to this model.
     center_of_mass : :class:`~compas.geometry.Point`
@@ -98,6 +100,16 @@ class TimberModel(Model):
             if isinstance(interaction, Joint):
                 joints.add(interaction)
         return joints
+
+    @property
+    def joint_candidates(self):
+        # type: () -> set[JointCandidate]
+        candidates = set()
+        for edge in self._graph.edges():
+            candidate = self._graph.edge_attribute(edge, "candidate")
+            if candidate is not None:
+                candidates.add(candidate)
+        return candidates
 
     @property
     def fasteners(self):
@@ -314,6 +326,49 @@ class TimberModel(Model):
             _ = self.add_interaction(element_a, element_b, joint)
             # TODO: should we create a bidirectional interaction here?
 
+    def add_joint_candidate(self, candidate):
+        # type: (JointCandidate) -> None
+        """Add a joint candidate to the model.
+
+        Joint candidates are stored on the graph edges under the "candidate" attribute,
+        separate from actual joints which are stored under the "interaction" attribute.
+
+        Parameters
+        ----------
+        candidate : :class:`~compas_timber.connections.JointCandidate`
+            An instance of a JointCandidate class.
+        """
+        for interaction in candidate.interactions:
+            element_a, element_b = interaction
+            edge = (element_a.graph_node, element_b.graph_node)
+            if edge not in self._graph.edges():
+                self._graph.add_edge(*edge)
+
+                # HACK: calls to `model.joints` expect there to be a "interactions" on any edges
+                self._graph.edge_attribute(edge, "interactions", [])
+
+            # this is how joints and candidates co-exist on the same edge, they are stored under different attributes
+            # (``interactions`` vs. ``candidate``)
+            # TODO: ``interactions`` is a list, should ``candidate`` be a list as well? don't see a reason rn.
+            self._graph.edge_attribute(edge, "candidate", candidate)
+
+    def remove_joint_candidate(self, candidate):
+        # type: (JointCandidate) -> None
+        """Removes this joint candidate from the model.
+
+        Parameters
+        ----------
+        candidate : :class:`~compas_timber.connections.JointCandidate`
+            The joint candidate to remove.
+        """
+        for interaction in candidate.interactions:
+            element_a, element_b = interaction
+            edge = (element_a.graph_node, element_b.graph_node)
+            if edge in self._graph.edges():
+                stored_candidate = self._graph.edge_attribute(edge, "candidate")
+                if stored_candidate is candidate:
+                    self._graph.unset_edge_attribute(edge, "candidate")
+
     def remove_joint(self, joint):
         # type: (Joint) -> None
         """Removes this joint object from the model.
@@ -381,7 +436,12 @@ class TimberModel(Model):
         return errors
 
     def connect_adjacent_beams(self, max_distance=None):
-        for joint in self.joints:
+        # Clear existing joint candidates
+        for candidate in list(self.joint_candidates):
+            self.remove_joint_candidate(candidate)
+
+        # Clear existing joints (except WallJoints)
+        for joint in list(self.joints):
             if not isinstance(joint, WallJoint):
                 self.remove_joint(joint)
 
@@ -401,7 +461,9 @@ class TimberModel(Model):
             p1, _ = intersection_line_line(beam_a.centerline, beam_b.centerline)
             p1 = Point(*p1) if p1 else None
 
-            JointCandidate.create(self, beam_a, beam_b, topology=topology, location=p1)
+            # Create candidate and add it to the model
+            candidate = JointCandidate(beam_a, beam_b, topology=topology, location=p1)
+            self.add_joint_candidate(candidate)
 
     def connect_adjacent_walls(self, max_distance=None):
         """Connects adjacent walls in the model.